Moved toast and modal functionality to their own files
